add metadata address info

+++ diancanrest/src/main/java/com/diancan/respositorys/RestaurantRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant,Integer> {
+public interface RestaurantRepository extends JpaRepository<Restaurant,Long> {
 
     // 通过餐厅地址模糊查询
     List<Restaurant> findRestaurantsByRestoAddressContains(String address);
